Fix popup_main bugs, refactor, update DA warning
The most crucial bug was that the fetching URL
was not updating along with the next/prev button
presses.

This was due to the fact the selected image did
not re-compute before sending.

This also adds some more type-safe defaults, and
refactors part of the code into their own
functions better.