Add repost detection
This uses the sha512 algorithm from crypto.subtle to generate
the hash that we can check on the furbooru.org API.

Additionally, this expands the warning system to display
multiple warnings. Given this increases the total vertical
space, and we're limited to 600px, we have to shrink
other parts of the addon.