Add an Options page
This is a major change, with a number of structural
reworks.

Now there's an options page, to cofigure your own
custom presets, stored to your local browser client.

Currently these settings aren't synced to your browser
user account, but I'm not sure I want to ever do that.
Certainly not as a default, but perhaps as an opt-in.

The options page currently is a pain to navigate
to, and it's really ugly.

The next steps are to add a little cog in the pop-up,
and make the page prettier.